Third Commit
Report: -created a tutorial sign. -created a sound system for both sfx and backgroundmusic. -created vsx for jumping and walking. -created a looping background. -changed the ui. tell me what do you think of it, it can still be improved. -created the boss. but i didnt animated it, because the animation i found wasnt that good and so i decided to show you just the idea which can then be replaced with other assets -created a plataform that once you touch it falls -created a settings menu but you can only change sound for now. -general fixes and improvements.